Update KSP tests
